refactor: organize Data Portal API handlers (#3556)
The changes here are to API handlers - their file paths, and their code sharing. This is almost entirely focused on reorganizing the Portal API, which was the most disorganized; the Curation API had largely followed the standard I implement here (though some tweaks were required), and the WMG and Gene Info API's are so small, and not really built around "resources" in any meaningful sense yet, that they are not worth messing with.

The actual standard is to put handlers in a file path structure that mirrors the API path, with the file itself being named after the last portion of the API path (A route of ../the/last_portion results in handler being in ../the/last_portion.py) EXCEPT in cases where subsequent directories would conflict in name with a handler file, such as when both ../this/path and also ../this/path/exist, in which case handlers for the former reside in ../this/path/actions.py.

All handler methods are named according to their HTTP method: GET, POST, etc.
