+++ backend/portal/api/app/v1/collections/collection_id/publish.py
