+++ backend/portal/api/curation/curation-api.yml
@@ -269,7 +269,7 @@ paths:
         All methods of updating or adding Datasets and metadata to a Revision work the same as for a private Collection.
 
         * Required: provide your access token in the request header as `Authorization: Bearer <access_token>`.
-      operationId: backend.portal.api.curation.v1.curation.collections.collection_id.revision.post_collection_revision
+      operationId: backend.portal.api.curation.v1.curation.collections.collection_id.revision.post
       security:
         - curatorAccess: []
       tags:
@@ -335,7 +335,7 @@ paths:
         the cancellation process is occurring, the Dataset may still appear as part of the Collection for a few minutes, but will eventually be removed.
 
         * Required: provide your access token in the request header as `Authorization: bearer <access_token>`.
-      operationId: backend.portal.api.curation.v1.curation.collections.collection_id.datasets.actions.delete
+      operationId: backend.portal.api.curation.v1.curation.collections.collection_id.datasets.dataset_id.actions.delete
       tags:
         - collection
       security:
@@ -364,7 +364,7 @@ paths:
       summary: Fetch Dataset metadata from a Collection.
       description: |
         Fetch Dataset metadata for `dataset_id` in a `collection_id`.
-      operationId: backend.portal.api.curation.v1.curation.collections.collection_id.datasets.actions.get
+      operationId: backend.portal.api.curation.v1.curation.collections.collection_id.datasets.dataset_id.actions.get
       tags:
         - collection
       parameters:
@@ -389,7 +389,7 @@ paths:
         file URL.
 
         * Required: provide your access token in the request header as `Authorization: Bearer <access_token>`.
-      operationId: backend.portal.api.curation.v1.curation.collections.collection_id.datasets.actions.put
+      operationId: backend.portal.api.curation.v1.curation.collections.collection_id.datasets.dataset_id.actions.put
       tags:
         - collection
       security:
@@ -476,7 +476,7 @@ paths:
         * Required: provide your access token in the request header as `Authorization: Bearer <access_token>`.
       security:
         - curatorAccess: []
-      operationId: backend.portal.api.curation.v1.curation.collections.collection_id.datasets.upload_s3.get
+      operationId: backend.portal.api.curation.v1.curation.collections.collection_id.s3_upload_credentials.get
       parameters:
         - $ref: "#/components/parameters/path_collection_id"
       responses:
@@ -516,7 +516,7 @@ paths:
         Generate Presigned S3 URLs to download the files associated with the `dataset_id`. 
         Pre-signed S3 URLs have a life span of up to 1 hour. You should
         generate a pre-signed URL immediately before each download to ensure it is current.
-      operationId: backend.portal.api.curation.v1.curation.collections.collection_id.assets.get
+      operationId: backend.portal.api.curation.v1.curation.collections.collection_id.datasets.dataset_id.assets.get
       parameters:
         - $ref: "#/components/parameters/path_collection_id"
         - $ref: "#/components/parameters/path_dataset_id"
