+++ backend/portal/api/curation/v1/curation/collections/collection_id/actions.py
@@ -1,11 +1,5 @@
 from flask import g, jsonify, Response
 
-from backend.portal.api.app.v1.collection import (
-    get_collection_and_verify_body,
-    get_publisher_metadata,
-    curation_get_normalized_doi_url,
-)
-
 from backend.api_server.db import dbconnect
 from backend.common.corpora_orm import (
     CollectionVisibility,
@@ -20,8 +14,17 @@ from backend.common.utils.http_exceptions import (
     ForbiddenHTTPException,
 )
 from backend.portal.api.app.v1.authorization import owner_or_allowed
-from backend.portal.api.app.v1.common import get_collection_else_forbidden
-from backend.portal.api.curation.v1.curation.collections.common import extract_doi_from_links, reshape_for_curation_api
+
+from backend.portal.api.collections_common import (
+    get_publisher_metadata,
+    get_collection_else_forbidden,
+    get_collection_and_verify_body,
+)
+from backend.portal.api.curation.v1.curation.collections.common import (
+    extract_doi_from_links,
+    reshape_for_curation_api,
+    curation_get_normalized_doi_url,
+)
 
 
 @dbconnect
