+++ backend/portal/api/curation/v1/curation/collections/collection_id/datasets/dataset_id/assets.py
@@ -4,7 +4,7 @@ from backend.api_server.db import dbconnect
 from backend.common.corpora_orm import DatasetArtifactFileType
 from backend.common.entities import DatasetAsset
 from backend.common.utils.http_exceptions import NotFoundHTTPException
-from backend.portal.api.app.v1.common import get_dataset_else_error
+from backend.portal.api.collections_common import get_dataset_else_error
 
 
 @dbconnect
