+++ backend/portal/api/curation/v1/curation/collections/collection_id/revision.py
@@ -1,10 +1,10 @@
 from flask import make_response, jsonify
 
 from backend.api_server.db import dbconnect
-from backend.portal.api.app.v1.collection import post_collection_revision_common
+from backend.portal.api.collections_common import post_collection_revision_common
 
 
 @dbconnect
-def post_collection_revision(collection_id: str, token_info: dict):
+def post(collection_id: str, token_info: dict):
     collection_revision = post_collection_revision_common(collection_id, token_info)
     return make_response(jsonify({"id": collection_revision.id}), 201)
