+++ tests/unit/backend/api_server/curation/collection/test_assets.py
@@ -59,7 +59,8 @@ class TestAsset(BaseAuthAPITest, CorporaTestCaseUsingMockAWS):
         self.assertEqual("Dataset not found.", actual_body["detail"])
 
     @patch(
-        "backend.portal.api.curation.v1.curation.collections.collection_id.assets.get_dataset_else_error",
+        "backend.portal.api.curation.v1.curation.collections.collection_id.datasets.dataset_id."
+        "assets.get_dataset_else_error",
         return_value=None,
     )
     def test__get_dataset_asset__asset_NOT_FOUND(self, get_dataset_else_error: Mock):
